airflow/requirements.txt
paho-mqtt>=2.1.0
apache-airflow-providers-openai>=1.7.2